LayoutTests/accessibility/mac/menu-item-values-expected.txt

This tests that the proper value is reported for menu item radio types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Item 1 Value: AXValue: 0
Item 1 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 2 Value: AXValue: 1
Item 2 Char mark: ✓
PASS item.isAttributeSupported('AXValue') is true


Item 3 Value: AXValue: 0
Item 3 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 4 Value: AXValue: 0
Item 4 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 5 Value: AXValue: 0
Item 5 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 6 Value: AXValue: 0
Item 6 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 7 Value: AXValue: 1
Item 7 Char mark: ✓
PASS item.isAttributeSupported('AXValue') is true


Item 8 Value: AXValue: 2
Item 8 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 9 Value: AXValue: 0
Item 9 Char mark:
PASS item.isAttributeSupported('AXValue') is true


Item 10 Value: AXValue: 0
Item 10 Char mark:
PASS item.isAttributeSupported('AXValue') is true


PASS successfullyParsed is true

TEST COMPLETE
